Convex classes of priors

The approach works for any convex class $\mathcal{G}$ of effect size distributions.

MinimaxCalibratedEBayes.ConvexPriorClass — Type

ConvexPriorClass

Abstract type representing convex classes of probability distributions $\mathcal{G}$.

Currently, the following classes are avalable:

MinimaxCalibratedEBayes.GaussianMixturePriorClass — Type

GaussianMixturePriorClass(σ_prior, grid, solver)

Class of distributions $G \in \mathcal{G}$ that can be written as the convolution of a Normal distribution with standard deviation σ_prior and a discrete distribution $\pi$ supported on grid.

\[G = \mathcal{N}(0,\sigma_{\text{prior}}^2) \star \pi\]

The solver object is a MathOptInterface compatible optimizer such as Mosek.Optimizer that will be used to solve the convex modulus of continuity (or worst-case-bias estimation) problem.

MinimaxCalibratedEBayes.HermitePriorClass — Type

HermitePriorClass(qmax, sobolev_order, sobolev_bound, solver)

Class of densities $g \in \mathcal{G}$ that can be written as $g(\mu) = \sum_{j=0}^{qmax} h_j(\mu)$, where $h_j$ is the $j$-th Hermite function and further satisfy the Sobolev constraint (with sobolev_order corresponding to $b$ and sobolev_bound to $C$):

\[\int_{-\infty}^{\infty} |g^*(t)|^2(t^2+1)^b dt \leq 2\pi C\]

The solver object is a MathOptInterface compatible optimizer such as Mosek.Optimizer that will be used to solve the convex modulus of continuity (or worst-case-bias estimation) problem.