Quasi-minimax linear estimators

Here we seek to demonstrate how we can solve optimization problems of the following form: Given a partition of the real line $(-\infty, \infty)$ into bins $I_{k}$, find the constant $Q_0$ and function $Q(\cdot)$ which optimize the following optimization problem:

\[\min_{Q_0,Q} \left\{ \sup_{ G \in \mathcal{G}} \{ (Q_0 + E_{G}[Q(Z_i)] - L(G))^2\} + \lambda(\delta)\frac{1}{n}\int Q^2(z) \bar{f}(z) dz \right\}\]

We solve this over all functions $Q(\cdot)$ that are piecewise constant on $I_k$. This corresponds to finding the linear estimator:

\[\hat{L} = Q_0 + \frac{1}{n} \sum_{i=1}^n Q(Z_i)\]

that solves a worst-case bias-variance problem, where the exact trade-off is parametrized by $\lambda(\delta)$.

Specification of the optimization problem consists of specifying the following:

The convex class of priors (effect size distributions) $\mathcal{G}$.

The linear functional$L(G)$ operating on $G \in \mathcal{G}$.

A pilot estimator for the marginal density of the observations $\bar{f}(\cdot)$.

A way of navigating the bias-variance tradeoff, i.e. of picking $\delta$.

MinimaxCalibratedEBayes.SteinMinimaxEstimator — Type

SteinMinimaxEstimator(Zs_discr::DiscretizedStandardNormalSamples,
                      prior_class::ConvexPriorClass,
                      target::LinearEBayesTarget,
                      δ_tuner::DeltaTuner

Computes a linear estimator optimizing a worst-case bias-variance tradeoff (specified by δ_tuner) for estimating a linear target over prior_class based on DiscretizedStandardNormalSamples Zs_discr.

Bias-variance tradeoff

The following options are available for navigating the bias-variance tradeoff:

MinimaxCalibratedEBayes.FixedDelta — Type

FixedDelta(δ) <: DeltaTuner

The simplest DeltaTuner. It just uses δ which has been fixed beforehand. This is used as a component of more complicated DeltaTuners which need to run repeatedly for different values of δ.

MinimaxCalibratedEBayes.RMSE — Type

RMSE(n::Integer, δ_min::Float64) <: DeltaTuner

A DeltaTuner that chooses the δ ≧ δ_min the optimizes the worst-case (root) mean squared error. Here n is the sample size used for estimation.

MinimaxCalibratedEBayes.HalfCIWidth — Type

HalfCIWidth(n::Integer, α::Float64 δ_min::Float64) <: DeltaTuner

A DeltaTuner that chooses the δ ≧ δ_min the optimizes the worst-case (root) mean squared error. Here n is the sample size used for estimation.

Worst case bias

The worst-case bias of an arbitrary linear estimator can be computed with the following function:

MinimaxCalibratedEBayes.worst_case_bias — Function

worst_case_bias(Q,
                Zs_discr::DiscretizedStandardNormalSamples,
                prior_class::ConvexPriorClass,
                target::LinearEBayesTarget)

Computes the worst-case positive and negative bias of the linear estimator Q for estimating the linear functional target over prior_class based on DiscretizedStandardNormalSamples Zs_discr.