Empirical Bayes estimands

In this section we introduce the interface for describing the inferential targets of the empirical Bayes analysis.

MinimaxCalibratedEBayes.EBayesTarget — Type

EBayesTarget

Abstract type that describe Empirical Bayes estimands (which we want to estimate or conduct inference for).

Any target::EBayesTarget may be used as a function on distributions, e.g., for a PriorDensityTarget

target = PriorDensityTarget(2.0)
target(Normal(0,1)) == pdf(Normal(0,1), 2.0)

true

Similarly for a MarginalDensityTarget

target = MarginalDensityTarget(StandardNormalSample(1.0))
target(Normal(0,1)) == pdf(Normal(0,sqrt(2)), 1.0)

true

Other interface functions implemented for all EBayesTargets include:

Base.extrema — Method

extrema(target::EBayesTarget)

Returns a tuple (a,b) such that a ≦ target ≦ b is always true. By default this is $(-\infty,+\infty)$, but e.g., for a PriorTailProbability it returns (0,1).

Linear Functionals

MinimaxCalibratedEBayes.LinearEBayesTarget — Type

LinearEBayesTarget <: EBayesTarget

Abstract type for Empirical Bayes estimands that are linear functionals of the prior $G$, i.e., they take the form $L(G)$ for some function linear functional $L$.

Interface

Distributions.cf — Method

cf(::LinearEBayesTarget, t)

The characteristic function of $L(\cdot)$, a LinearEBayesTarget, which we define as follows:

For $L(\cdot)$ which may be written as $L(G) = \int \psi(\mu)dG\mu$ (for a measurable function $\psi$) this returns the Fourier transform of $\psi$ evaluated at t, i.e., $\psi^*(t) = \int \exp(it x)\psi(x)dx$.

Note that $\psi^*(t)$ is such that for distributions $G$ with density $g$ (and $g^*$ the Fourier Transform of $g$) the following holds:

\[L(G) = \frac{1}{2\pi}\int g^*(\mu)\psi^*(\mu) d\mu\]

Implemented linear targets

MinimaxCalibratedEBayes.MarginalDensityTarget — Type

MarginalDensityTarget(Z::StandardNormalSample) <: LinearEBayesTarget

Example call

MarginalDensityTaget(StandardNormalSample(2.0))

Description

Describes the marginal density evaluated at $Z=z$ (e.g. $Z=2$ in the example above) of a sample drawn from the hierarchical model

\[\mu \sim G, Z \sim \mathcal{N}(0,1)\]

In other words, letting $\phi$ the Standard Normal pdf

\[L(G) = \phi \star dG(z)\]

Note that 2.0 has to be wrapped inside StandardNormalSample(2.0) since this target depends not only on G and the location, but also on the likelihood. Additional likelihoods will be added in the future.

MinimaxCalibratedEBayes.PriorDensityTarget — Type

PriorDensityTarget(z::Float64) <: LinearEBayesTarget

Example call

PriorDensityTarget(2.0)

Description

This is the evaluation functional of the density of $G$ at z, i.e., $L(G) = G'(z) = g(z)$ or in Julia code L(G) = pdf(G, z).

MinimaxCalibratedEBayes.PriorTailProbability — Type

PriorTailProbability(cutoff::Float64) <: LinearEBayesTarget

Example call

PriorTailProbability(2.0)

Description

This is the evaluation functional of the tail probability of $G$ at cutoff, i.e., $L(G) = 1-G(\text{cutoff})$ or in Julia code L(G) = ccdf(G, z).

Posterior Estimands