Pilot estimators

Butucea-Comte

MinimaxCalibratedEBayes.ButuceaComteOptions — Type

ButuceaComteOptions(; bandwidth = :auto)

The Butucea Comte estimator of linear functional in the deconvolution model. Given a LinearEBayesTarget with characteristic function $\psi^*$ and samples convolved with Gaussian noise (with $\phi$ Standard Gaussian pdf), it is defined as follows

\[\hat{L}_{\text{BC},h_n} = \frac{1}{2 \pi n}\sum_{i=1}^n \int_{-1/h_n}^{1/h_n} \exp(it Z_k) \frac{\psi^*(-t)}{\varphi^*(t)}dt\]

$h$ is the bandwidth, the option :auto will pick it automatically.

Reference:

Butucea, C. and Comte, F., 2009. Adaptive estimation of linear functionals in the convolution model and applications. Bernoulli, 15(1), pp.69-98.

A target::LinearEBayesTarget can be estiamted based on Standard Normal samples Zs with the Butucea-Comte method bcopt::ButuceaComteOptions as follows

estimate(target, bcop, Zs)

Distributions.estimate — Method

estimate(target::LinearEBayesTarget, bcopt::ButuceaComteOptions, Zs)

bla