Discretization and Localization

Empirical Bayes samples

Discretization

MinimaxCalibratedEBayes.DiscretizedStandardNormalSamples — Type

DiscretizedStandardNormalSamples(marginal_grid)

This type represents discretization of Standard Normal Samples to the grid marginal_grid. Let $t_1 < \dotsc, < t_K$ the elements of marginal_grid, then this type implies that we map each sample $Z \sim \mathcal{N}(0,1)$ to one of the intervals $[t_j, t_{j+1}),j=0,\dotsc,K$, with $t_0 = -\infty$ and $t_{k+1} = +\infty$. Our inference will depend only on the discretized samples.

DiscretizedStandardNormalSamples(Zs::AbstractVector{<:StandardNormalSample}, marginal_grid)

If we also provide the constructor with a vector of StandardNormalSample's, then it also stores the Histogram of the samples (according to the binning defined above) in the mhist slot. This is the multinomial sufficient statistics of the discretized data.

The DiscretizedStandardNormalSamples type also contains three more slots:

f_min,f_max: These can be either nothing or vectors of length equal to the number of

histogram bins. They contain bounds so that:

\[f_{\text{min}}[k] \leq \mathbb P_G[ Z \in \text{bin }k] \leq f_{\text{max}}[k] \]

var_proxy: This is typically a vector of point estimates of $\mathbb P_G[ Z \in \text{bin }k]$.

Localization

Neighborhood construction

MinimaxCalibratedEBayes.KDEInfinityBandOptions — Type

KDEInfinityBandOptions(; kernel=DeLaValleePoussinKernel
                         bandwidth = nothing,
                         a_min, 
                         a_max,
                         nboot = 1000)

This struct contains hyperparameters that will be used for constructing a neighborhood of the marginal density. The steps of the method (and corresponding hyperparameter meanings) are as follows

First a kernel density estimate $\bar{f}$ of the data is fit with kernel as the

kernel and bandwidth (the default bandwidth = nothing corresponds to automatic bandwidth selection).

Second, a Poisson bootstrap with nboot replication will be used to estimate a $L_{\infty}$

neighborhood $c_m$ of the true density $f$ which is such that with probability tending to 1:

\[\sup_{x \in [a_{\text{min}} , a_{\text{max}}]} | \bar{f}(x) - f(x)| \leq c_m\]

Note that the bound is valid from a_min to a_max.

Reference:

Paul Deheuvels and Gérard Derzko. Asymptotic certainty bands for kernel density estimators based upon a bootstrap resampling scheme. In Statistical models and methods for biomedical and technical systems, pages 171–186. Springer, 2008

Kernels that can be used:

MinimaxCalibratedEBayes.DeLaValleePoussinKernel — Type

DeLaValleePoussinKernel(h)

Implements the DeLaValleePoussinKernel with bandwidth h to be used for kernel density estimation through the KernelDensity.jl package. The De La Vallée-Poussin kernel is defined as follows:

\[K_V(x) = \frac{\cos(x)-\cos(2x)}{\pi x^2}\]

Its use case is similar to the SincKernel, however it has the advantage of being integrable (in the Lebesgue sense). Its Fourier transform is the following:

\[K^*_V(t) = \begin{cases} 1, & \text{ if } x\in[-1,1] \\ 0, &\text{ if } |t| \geq 2 \\ 2-|t|,& \text{ if } |t| \in [1,2] \end{cases}\]

MinimaxCalibratedEBayes.SincKernel — Type

SincKernel(h)

Implements the SincKernel with bandwidth h to be used for kernel density estimation through the KernelDensity.jl package. The sinc kernel is defined as follows:

\[K_{\text{sinc}}(x) = \frac{\sin(x)}{\pi x} \]

It is not typically used for kernel density estimation, because this kernel is not a density itself. However, it is particularly well suited to deconvolution problems and estimation of very smooth densities because its Fourier transform is the following:

\[K^*_{\text{sinc}}(t) = \mathbf 1( t \in [-1,1])\]